Refactor - moved all meta data function to one file
