+++ govcd/metadata.go
@@ -1,0 +1,107 @@
+/*
+ * Copyright 2019 VMware, Inc.  All rights reserved.  Licensed under the Apache v2 License.
+ */
+
+package govcd
+
+import (
+	"github.com/vmware/go-vcloud-director/v2/types/v56"
+	"net/http"
+	"net/url"
+	"strings"
+)
+
+// GetMetadata() function calls private function getMetadata() with vm.client and vm.VM.HREF
+// which returns a *types.Metadata struct for provided VM input.
+func (vm *VM) GetMetadata() (*types.Metadata, error) {
+	return getMetadata(vm.client, vm.VM.HREF)
+}
+
+// DeleteMetadata() function calls private function deleteMetadata() with vm.client and vm.VM.HREF
+// which deletes metadata depending on key provided as input from VM.
+func (vm *VM) DeleteMetadata(key string) (Task, error) {
+	return deleteMetadata(vm.client, key, vm.VM.HREF)
+}
+
+// AddMetadata() function calls private function addMetadata() with vm.client and vm.VM.HREF
+// which adds metadata key, value pair provided as input to VM.
+func (vm *VM) AddMetadata(key string, value string) (Task, error) {
+	return addMetadata(vm.client, key, value, vm.VM.HREF)
+}
+
+// GetMetadata() function returns meta data for vDC.
+func (vdc *Vdc) GetMetadata() (*types.Metadata, error) {
+	return getMetadata(vdc.client, getAdminVdcURL(vdc.Vdc.HREF))
+}
+
+// DeleteMetadata() function deletes metadata by key provided as input
+func (vdc *Vdc) DeleteMetadata(key string) (Task, error) {
+	return deleteMetadata(vdc.client, key, getAdminVdcURL(vdc.Vdc.HREF))
+}
+
+// AddMetadata() function adds metadata key, value pair provided as input to vDC.
+func (vdc *Vdc) AddMetadata(key string, value string) (Task, error) {
+	return addMetadata(vdc.client, key, value, getAdminVdcURL(vdc.Vdc.HREF))
+}
+
+func getAdminVdcURL(vdcURL string) string {
+	return strings.Split(vdcURL, "/api/vdc/")[0] + "/api/admin/vdc/" + strings.Split(vdcURL, "/api/vdc/")[1]
+}
+
+// GetMetadata() function calls private function getMetadata() with vapp.client and vapp.VApp.HREF
+// which returns a *types.Metadata struct for provided vapp input.
+func (vapp *VApp) GetMetadata() (*types.Metadata, error) {
+	return getMetadata(vapp.client, vapp.VApp.HREF)
+}
+
+func getMetadata(client *Client, requestUri string) (*types.Metadata, error) {
+	metadata := &types.Metadata{}
+
+	_, err := client.ExecuteRequest(requestUri+"/metadata/", http.MethodGet,
+		types.MimeMetaData, "error retrieving metadata: %s", nil, metadata)
+
+	return metadata, err
+}
+
+// DeleteMetadata() function calls private function deleteMetadata() with vapp.client and vapp.VApp.HREF
+// which deletes metadata depending on key provided as input from vApp.
+func (vapp *VApp) DeleteMetadata(key string) (Task, error) {
+	return deleteMetadata(vapp.client, key, vapp.VApp.HREF)
+}
+
+// Deletes metadata (type MetadataStringValue) from the vApp
+// TODO: Support all MetadataTypedValue types with this function
+func deleteMetadata(client *Client, key string, requestUri string) (Task, error) {
+	apiEndpoint, _ := url.ParseRequestURI(requestUri)
+	apiEndpoint.Path += "/metadata/" + key
+
+	// Return the task
+	return client.ExecuteTaskRequest(apiEndpoint.String(), http.MethodDelete,
+		"", "error deleting metadata: %s", nil)
+}
+
+// AddMetadata() function calls private function addMetadata() with vapp.client and vapp.VApp.HREF
+// which adds metadata key, value pair provided as input.
+func (vapp *VApp) AddMetadata(key string, value string) (Task, error) {
+	return addMetadata(vapp.client, key, value, vapp.VApp.HREF)
+}
+
+// Adds metadata (type MetadataStringValue) to the vApp
+// TODO: Support all MetadataTypedValue types with this function
+func addMetadata(client *Client, key string, value string, requestUri string) (Task, error) {
+	newMetadata := &types.MetadataValue{
+		Xmlns: types.XMLNamespaceVCloud,
+		Xsi:   types.XMLNamespaceXSI,
+		TypedValue: &types.TypedValue{
+			XsiType: "MetadataStringValue",
+			Value:   value,
+		},
+	}
+
+	apiEndpoint, _ := url.ParseRequestURI(requestUri)
+	apiEndpoint.Path += "/metadata/" + key
+
+	// Return the task
+	return client.ExecuteTaskRequest(apiEndpoint.String(), http.MethodPut,
+		types.MimeMetaDataValue, "error adding metadata: %s", newMetadata)
+}
