+++ govcd/vdc_test.go
@@ -236,53 +236,3 @@ func (vcd *TestVCD) Test_QueryVM(check *C) {
 func init() {
 	testingTags["vdc"] = "vdc_test.go"
 }
-
-func (vcd *TestVCD) Test_AddMetadataForVdc(check *C) {
-	if vcd.config.VCD.Vdc == "" {
-		check.Skip("skipping test because no vDC name is empty")
-	}
-
-	fmt.Printf("Running: %s\n", check.TestName())
-
-	// Add metadata
-	task, err := vcd.vdc.AddMetadata("key", "value")
-	check.Assert(err, IsNil)
-	err = task.WaitTaskCompletion()
-	check.Assert(err, IsNil)
-	check.Assert(task.Task.Status, Equals, "success")
-
-	// Check if metadata was added correctly
-	metadata, err := vcd.vdc.GetMetadata()
-	check.Assert(err, IsNil)
-	check.Assert(metadata.MetadataEntry[0].Key, Equals, "key")
-	check.Assert(metadata.MetadataEntry[0].TypedValue.Value, Equals, "value")
-}
-
-func (vcd *TestVCD) Test_DeleteMetadataForVdc(check *C) {
-	if vcd.config.VCD.Vdc == "" {
-		check.Skip("skipping test because no vDC name is empty")
-	}
-
-	fmt.Printf("Running: %s\n", check.TestName())
-
-	// Add metadata
-	task, err := vcd.vdc.AddMetadata("key2", "value2")
-	check.Assert(err, IsNil)
-	err = task.WaitTaskCompletion()
-	check.Assert(err, IsNil)
-	check.Assert(task.Task.Status, Equals, "success")
-
-	// Remove metadata
-	task, err = vcd.vdc.DeleteMetadata("key2")
-	check.Assert(err, IsNil)
-	err = task.WaitTaskCompletion()
-	check.Assert(err, IsNil)
-	check.Assert(task.Task.Status, Equals, "success")
-	metadata, err := vcd.vdc.GetMetadata()
-	check.Assert(err, IsNil)
-	for _, k := range metadata.MetadataEntry {
-		if k.Key == "key2" {
-			check.Errorf("metadata.MetadataEntry should not contain key: %s", k)
-		}
-	}
-}
