+++ govcd/vm.go
@@ -553,24 +553,6 @@ func (vm *VM) GetQuestion() (types.VmPendingQuestion, error) {
 
 }
 
-// GetMetadata() function calls private function getMetadata() with vm.client and vm.VM.HREF
-// which returns a *types.Metadata struct for provided VM input.
-func (vm *VM) GetMetadata() (*types.Metadata, error) {
-	return getMetadata(vm.client, vm.VM.HREF)
-}
-
-// DeleteMetadata() function calls private function deleteMetadata() with vm.client and vm.VM.HREF
-// which deletes metadata depending on key provided as input from VM.
-func (vm *VM) DeleteMetadata(key string) (Task, error) {
-	return deleteMetadata(vm.client, key, vm.VM.HREF)
-}
-
-// AddMetadata() function calls private function addMetadata() with vm.client and vm.VM.HREF
-// which adds metadata key, value pair provided as input to VM.
-func (vm *VM) AddMetadata(key string, value string) (Task, error) {
-	return addMetadata(vm.client, key, value, vm.VM.HREF)
-}
-
 // Use the provide answer to existing VM question for operation which need additional response
 // Reference:
 // https://code.vmware.com/apis/287/vcloud#/doc/doc/operations/POST-AnswerVmPendingQuestion.html
